feat: Add dark mode support with persistent theme preference
feat: Add dark mode support with persistent theme preference

Changes made:
1. Frontend Updates:
   - Added theme toggle button with sun/moon icons in navigation
   - Implemented CSS variables for theming
   - Added dark theme color scheme
   - Added theme persistence using localStorage

2. CSS Changes:
   - Created theme variables for colors and shadows
   - Updated all color values to use CSS variables
   - Added dark mode color scheme under [data-theme="dark"]
   - Added theme toggle button styles

3. HTML Changes:
   - Added theme toggle button in navigation bar
   - Added aria-label for accessibility
   - Added theme icon and text elements

4. JavaScript Changes:
   - Added theme management functions
   - Implemented theme toggle functionality
   - Added theme persistence in localStorage
   - Added theme initialization on page load

5. Backend Cleanup:
   - Consolidated middleware code
   - Removed unused metrics functionality
   - Simplified server configuration

Technical Details:
- Uses CSS custom properties for theme colors
- Theme preference persists across page reloads
- Smooth transition between themes
- Accessibility-friendly implementation

Test Steps:
1. Light/Dark mode toggle works
2. Theme persists after page refresh
3. All UI elements respect theme colors
4. No console errors
5. Proper icon transitions (🌞 ↔ 🌙)
