--- internal/middleware/logger.go
@@ -1,44 +1,0 @@
-package middleware
-
-import (
-	"log/slog"
-	"net/http"
-	"time"
-)
-
-func RequestLoggerMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		// Create a custom response writer to capture the status code
-		rw := &responseWriter{
-			ResponseWriter: w,
-			statusCode:    http.StatusOK,
-		}
-
-		next.ServeHTTP(rw, r)
-
-		duration := time.Since(start)
-
-		// Log request details with structured logging
-		slog.Info("request completed",
-			"method", r.Method,
-			"path", r.URL.Path,
-			"status", rw.statusCode,
-			"duration", duration,
-			"remote_addr", r.RemoteAddr,
-			"user_agent", r.UserAgent(),
-		)
-	})
-}
-
-// responseWriter is a custom response writer that captures the status code
-type responseWriter struct {
-	http.ResponseWriter
-	statusCode int
-}
-
-func (rw *responseWriter) WriteHeader(code int) {
-	rw.statusCode = code
-	rw.ResponseWriter.WriteHeader(code)
-}
