+++ public/index.html
@@ -17,6 +17,12 @@
                 <li><a href="#features">Features</a></li>
                 <li><a href="#api">API</a></li>
                 <li><a href="#docs">Docs</a></li>
+                <li>
+                    <button id="theme-toggle" class="theme-toggle" aria-label="Toggle dark mode">
+                        <span id="theme-toggle-icon">🌞</span>
+                        <span id="theme-toggle-text">Light</span>
+                    </button>
+                </li>
             </ul>
         </nav>
     </header>
