+++ public/styles.css
@@ -1,3 +1,30 @@
+/* Theme Variables */
+:root {
+    /* Light theme (default) */
+    --bg-primary: #fff;
+    --bg-secondary: #f8f9fa;
+    --text-primary: #333;
+    --text-secondary: #666;
+    --card-bg: #f8f9fa;
+    --card-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    --code-bg: #f1f3f4;
+    --response-bg: #2d3748;
+    --response-text: #a0aec0;
+}
+
+[data-theme="dark"] {
+    /* Dark theme */
+    --bg-primary: #1a1a1a;
+    --bg-secondary: #2d2d2d;
+    --text-primary: #e0e0e0;
+    --text-secondary: #b0b0b0;
+    --card-bg: #2d2d2d;
+    --card-shadow: 0 4px 6px rgba(0, 0, 0, 0.3);
+    --code-bg: #3a3a3a;
+    --response-bg: #1a1a1a;
+    --response-text: #b0b0b0;
+}
+
 /* FeatherJet Demo Styles */
 * {
     margin: 0;
@@ -8,7 +35,7 @@
 body {
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
     line-height: 1.6;
-    color: #333;
+    color: var(--text-primary);
     background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
     min-height: 100vh;
 }
@@ -131,7 +158,7 @@ nav {
 
 /* Features Section */
 .features {
-    background: white;
+    background: var(--bg-primary);
     padding: 4rem 0;
     margin-top: 2rem;
 }
@@ -140,7 +167,7 @@ nav {
     text-align: center;
     font-size: 2.5rem;
     margin-bottom: 3rem;
-    color: #333;
+    color: var(--text-primary);
 }
 
 .feature-grid {
@@ -150,10 +177,10 @@ nav {
 }
 
 .feature-card {
-    background: #f8f9fa;
+    background: var(--card-bg);
     padding: 2rem;
     border-radius: 12px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    box-shadow: var(--card-shadow);
     transition: transform 0.3s;
 }
 
@@ -174,7 +201,7 @@ nav {
 
 /* API Section */
 .api-section {
-    background: #f8f9fa;
+    background: var(--bg-secondary);
     padding: 4rem 0;
 }
 
@@ -182,7 +209,7 @@ nav {
     text-align: center;
     font-size: 2.5rem;
     margin-bottom: 3rem;
-    color: #333;
+    color: var(--text-primary);
 }
 
 .api-list {
@@ -195,10 +222,10 @@ nav {
     align-items: center;
     gap: 1rem;
     padding: 1rem;
-    background: white;
+    background: var(--bg-primary);
     border-radius: 8px;
     margin-bottom: 1rem;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    box-shadow: var(--card-shadow);
 }
 
 .method {
